refactor sep logic
