+++ src/services/anchor/sep10/utils.ts
@@ -1,0 +1,77 @@
+import { Keyring } from '@polkadot/api';
+import { keccak256 } from 'viem/utils';
+
+import { fetchSep10Signatures, SignerServiceSep10Request } from '../../signingService';
+import { SIGNING_SERVICE_URL } from '../../../constants/constants';
+import { config } from '../../../config';
+
+// Returns the hash value for the address.
+// If it's a polkadot address, it will return raw data of the address.
+function getHashValueForAddress(address: string) {
+  if (address.startsWith('0x')) {
+    return address as `0x${string}`;
+  } else {
+    const keyring = new Keyring({ type: 'sr25519' });
+    return keyring.decodeAddress(address);
+  }
+}
+
+// A memo derivation.
+async function deriveMemoFromAddress(address: string) {
+  const hashValue = getHashValueForAddress(address);
+  const hash = keccak256(hashValue);
+  return BigInt(hash).toString().slice(0, 15);
+}
+
+export const exists = (value?: string | null): value is string => !!value && value?.length > 0;
+
+export async function sep10SignaturesWithLoginRefresh(
+  refreshFunction: () => Promise<void>,
+  args: SignerServiceSep10Request,
+) {
+  try {
+    return await fetchSep10Signatures(args);
+  } catch (error: unknown) {
+    if (error instanceof Error && error.message === 'Invalid signature') {
+      await refreshFunction();
+      return await fetchSep10Signatures(args);
+    }
+    throw new Error('Could not fetch sep 10 signatures from backend');
+  }
+}
+
+// Return the URLSearchParams and the account (master/omnibus or ephemeral) that was used for SEP-10
+export async function getUrlParams(
+  ephemeralAccount: string,
+  usesMemo: boolean,
+  supportsClientDomain: boolean,
+  address: string,
+): Promise<{ urlParams: URLSearchParams; sep10Account: string }> {
+  let sep10Account: string;
+  const params = new URLSearchParams();
+
+  if (usesMemo) {
+    const response = await fetch(`${SIGNING_SERVICE_URL}/v1/stellar/sep10`);
+    if (!response.ok) {
+      throw new Error('Failed to fetch client master SEP-10 public account.');
+    }
+
+    const { masterSep10Public } = await response.json();
+    if (!masterSep10Public) {
+      throw new Error('masterSep10Public not found in response.');
+    }
+
+    sep10Account = masterSep10Public;
+    params.append('account', sep10Account);
+    params.append('memo', await deriveMemoFromAddress(address));
+  } else {
+    sep10Account = ephemeralAccount;
+    params.append('account', sep10Account);
+  }
+
+  if (supportsClientDomain) {
+    params.append('client_domain', config.applicationClientDomain);
+  }
+
+  return { urlParams: params, sep10Account };
+}
