+++ src/services/offrampingFlow.ts
@@ -10,7 +10,7 @@ import { u8aToHex } from '@polkadot/util';
 import { SigningPhase } from '../hooks/offramp/useMainProcess';
 import { TrackableEvent } from '../contexts/events';
 import { isNetworkEVM, Networks } from '../contexts/network';
-import { SepResult } from './anchor';
+import { SepResult } from '../types/sep';
 
 import {
   getInputTokenDetailsOrDefault,
