+++ src/types/sep.ts
@@ -1,0 +1,28 @@
+import { OutputTokenDetails } from '../constants/tokenConfig';
+
+export interface TomlValues {
+  signingKey?: string;
+  webAuthEndpoint?: string;
+  sep24Url?: string;
+  sep6Url?: string;
+  kycServer?: string;
+}
+
+export interface ISep24Intermediate {
+  url: string;
+  id: string;
+}
+
+export interface IAnchorSessionParams {
+  token: string;
+  tomlValues: TomlValues;
+  tokenConfig: OutputTokenDetails;
+  offrampAmount: string;
+}
+
+export interface SepResult {
+  amount: string;
+  memo: string;
+  memoType: string;
+  offrampingAccount: string;
+}
